+++ app/controllers/application_controller.rb
@@ -61,7 +61,7 @@ class ApplicationController < ActionController::Base
   def configure_permitted_parameters
     devise_parameter_sanitizer.permit(
       :sign_up,
-      keys: [:full_name, :email, :password]
+      keys: [:full_name, :email, :email_confirmation, :password]
     )
   end